+++ packages/core/src/components.d.ts
@@ -19,7 +19,6 @@ import { ListButtonChangeEvent, ListButtonSelectedEvent } from "./components/lis
 import { BaseLayer, BaseLayerChangeEvent } from "./components/map-base-layers/map-base-layers.interfaces";
 import { MapControlsToggleEvent } from "./components/map-controls/map-controls.interfaces";
 import { Overlay, OverlayChangeEvent } from "./components/map-overlays/map-overlays.interfaces";
-import { ModalRole } from "./components/modal/modal";
 import { DsoModalCloseEvent } from "./components/modal/modal.interfaces";
 import { OzonContentAnchorClick, OzonContentClick } from "./components/ozon-content/ozon-content.interfaces";
 import { PaginationSelectPageEvent } from "./components/pagination/pagination.interfaces";
@@ -325,7 +324,7 @@ export namespace Components {
         /**
           * the role for the modal `dialog` | `alert` | `alertdialog` defaults to `dialog`
          */
-        "role": ModalRole;
+        "role": string;
         /**
           * when `false` the close button in the header will not be rendered. Defaults to `true`
          */
@@ -1173,7 +1172,7 @@ declare namespace LocalJSX {
         /**
           * the role for the modal `dialog` | `alert` | `alertdialog` defaults to `dialog`
          */
-        "role"?: ModalRole;
+        "role"?: string;
         /**
           * when `false` the close button in the header will not be rendered. Defaults to `true`
          */